Refactor slate editor component
