+++ src/App.css
@@ -7,26 +7,3 @@
     margin: 65px auto 45px;
     border-radius: 4.5px;
 }
-
-.format-toolbar {
-    display: flex;
-    border-bottom: solid 1.7px rgba(199, 198, 255, 0.15);
-    padding: 10px 0;
-    margin: 0 0 10px 0;
-}
-
-.tooltip-icon-button {
-    border: 0;
-    margin: 0 5px 0 0;
-}
-
-.tooltip-icon-button:hover {
-    cursor: pointer;
-}
-
-blockquote {
-    background: #f9f9f9;
-    border-left: 4px solid #ccc;
-    margin: 1.5em 10px;
-    padding: 0.5em 10px;
-}
