--- src/components/Toolbar.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-const Toolbar = props => <div className="format-toolbar">{props.children}</div>;
-
-export default Toolbar;
