+++ src/components/editor/TextEditor.css
@@ -1,0 +1,22 @@
+.format-toolbar {
+  display: flex;
+  border-bottom: solid 1.7px rgba(199, 198, 255, 0.15);
+  padding: 10px 0;
+  margin: 0 0 10px 0;
+}
+
+.tooltip-icon-button {
+  border: 0;
+  margin: 0 5px 0 0;
+}
+
+.tooltip-icon-button:hover {
+  cursor: pointer;
+}
+
+blockquote {
+  background: #f9f9f9;
+  border-left: 4px solid #ccc;
+  margin: 1.5em 10px;
+  padding: 0.5em 10px;
+}
