+++ src/components/editor/blocks/index.js
@@ -1,0 +1,15 @@
+import BlockQuote from "./BlockQuote";
+import BulletedList from "./BulletedList";
+import HeadingOne from "./HeadingOne";
+import HeadingTwo from "./HeadingTwo";
+import ListItem from "./ListItem";
+import NumberedList from "./NumberedList";
+
+export {
+  BlockQuote,
+  BulletedList,
+  HeadingOne,
+  HeadingTwo,
+  ListItem,
+  NumberedList
+};
