+++ src/components/editor/initialValue.js
@@ -1,0 +1,22 @@
+import { Value } from "slate";
+
+export default Value.fromJSON({
+  document: {
+    nodes: [
+      {
+        object: "block",
+        type: "paragraph",
+        nodes: [
+          {
+            object: "text",
+            leaves: [
+              {
+                text: ""
+              }
+            ]
+          }
+        ]
+      }
+    ]
+  }
+});
