+++ src/components/editor/marks/index.js
@@ -1,0 +1,6 @@
+import Bold from "./Bold";
+import Code from "./Code";
+import Italic from "./Italic";
+import Underline from "./Underline";
+
+export { Bold, Code, Italic, Underline };
