+++ src/components/editor/plugins/index.js
@@ -1,0 +1,3 @@
+import MarkHotKey from "./MarkHotKey";
+
+export { MarkHotKey };
