+++ src/components/editor/toolbar/icons/index.js
@@ -1,0 +1,21 @@
+import { ic_format_bold } from "react-icons-kit/md/ic_format_bold";
+import { ic_format_italic } from "react-icons-kit/md/ic_format_italic";
+import { ic_format_underlined } from "react-icons-kit/md/ic_format_underlined";
+import { ic_code } from "react-icons-kit/md/ic_code";
+import { ic_format_quote } from "react-icons-kit/md/ic_format_quote";
+import { ic_format_list_bulleted } from "react-icons-kit/md/ic_format_list_bulleted";
+import { ic_format_list_numbered } from "react-icons-kit/md/ic_format_list_numbered";
+import { ic_looks_one } from "react-icons-kit/md/ic_looks_one";
+import { ic_looks_two } from "react-icons-kit/md/ic_looks_two";
+
+export {
+  ic_format_bold,
+  ic_format_italic,
+  ic_format_underlined,
+  ic_code,
+  ic_format_quote,
+  ic_format_list_bulleted,
+  ic_format_list_numbered,
+  ic_looks_one,
+  ic_looks_two
+};
