--- src/components/index.js
@@ -1,27 +1,0 @@
-import TextEditor from "./TextEditor";
-import BoldMark from "./marks/BoldMark";
-import ItalicMark from "./marks/ItalicMark";
-import UnderlineMark from "./marks/UnderlineMark";
-import CodeMark from "./marks/CodeMark";
-import BlockQuote from "./blocks/BlockQuote";
-import BulletedList from "./blocks/BulletedList";
-import NumberedList from "./blocks/NumberedList";
-import ListItem from "./blocks/ListItem";
-import HeadingOne from "./blocks/HeadingOne";
-import HeadingTwo from "./blocks/HeadingTwo";
-import Toolbar from "./Toolbar";
-
-export {
-  TextEditor,
-  BoldMark,
-  ItalicMark,
-  CodeMark,
-  BlockQuote,
-  BulletedList,
-  NumberedList,
-  ListItem,
-  HeadingOne,
-  HeadingTwo,
-  UnderlineMark,
-  Toolbar
-};
